pyproject.toml
[project]
name = "netrias-client"
version = "0.0.1"
description = "Ergonomic Python client for the Netrias harmonization API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0",
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=7",
    "pytest-asyncio>=0.23",
    "ruff>=0.5.0",
    "basedpyright",
    "typing-extensions",
    "ty",
    "python-dotenv>=1.0",
]

[tool.ruff]
unsafe-fixes = false
line-length = 140
select = [
    "E", # Pycodestyle errors
    "F", # Pyflakes
    "I", # Import order
    "C", # Pylint convention messages
    "R", # Pylint refactor messages
    "W", # Pylint warning messages
    "E", # Pylint error messages
]
ignore = [
    "F401", # Ignore unused import warning
    "F601", # Ignore dictionary key literal repeated warning
    "F811", # Ignore redefinition of unused name warning
    "I001", # Ignore import block is un-sorted or un-formatted
    "W293", # Ignore whitespace warning
    "W291", # Ignore trailing whitespace warning
    "E712", # Ignore comparisons to True/False directly
]

[tool.ruff.lint.mccabe]
max-complexity = 4

[tool.ruff.lint]
select = [
    "E", # Pycodestyle errors
    "F", # Pyflakes
    "I", # Import order
    "C", # Pylint convention messages
    "R", # Pylint refactor messages
    "W", # Pylint warning messages
    "E", # Pylint error messages
]
ignore = [
    "F401", # Ignore unused import warning
    "F601", # Ignore dictionary key literal repeated warning
    "F811", # Ignore redefinition of unused name warning
    "I001", # Ignore import block is un-sorted or un-formatted
    "W293", # Ignore whitespace warning
    "W291", # Ignore trailing whitespace warning
    "E712", # Ignore comparisons to True/False directly
]

[tool.ruff.lint.per-file-ignores]
"tests/**.py" = ["E501"]
